# SYS-1891: Improve export performance