--- a/ftva_lab_data/templates/release_notes.html
+++ b/ftva_lab_data/templates/release_notes.html
@@ -4,6 +4,20 @@
 <h3>Release Notes</h3>
 <hr/>
 
+<h4>1.0.2</h4>
+<p><i>July 3, 2025</i></p>
+<ul>
+    <li>Exports are now in CSV instead of XLSX.</li>
+    <li>Exports are significantly faster.</li>
+    <li>The spinner / download indicator is now active only during export.</li>
+</ul>
+
+<h4>1.0.1</h4>
+<p><i>July 1, 2025</i></p>
+<ul>
+    <li>Tweaked timeout for long exports.</li>
+</ul>
+
 <h4>1.0.0</h4>
 <p><i>July 1, 2025</i></p>
 <ul>