--- a/ftva_lab_data/templates/search_results.html
+++ b/ftva_lab_data/templates/search_results.html
@@ -50,15 +50,6 @@
         <button type="submit" class="btn btn-sm btn-primary">Assign Selected</button>
     </form>
     {% endif %}
-    <form method="get" action="{% url 'export_search_results' %}" class="mb-3" id="export-form">
-        {% csrf_token %}
-        <input type="hidden" name="search" value="{{ search }}">
-        <input type="hidden" name="search_column" value="{{ search_column }}">
-        <button type="button" class="btn btn-outline-success" id="export-button">
-            Export to CSV
-        </button>
-    </form>
-</div> 
 
     <div class="mb-3">
         <button
@@ -66,7 +57,7 @@
             class="btn btn-outline-success h-100" 
             id="export-button"
             hx-on:click="handleExportSearchResults(this, event)"
-            >Export to Excel
+            >Export to CSV
                 <span id="export-spinner"></span>
         </button>
     </div>