--- a/ftva_lab_data/views.py
+++ b/ftva_lab_data/views.py
@@ -288,6 +288,8 @@ def export_search_results(request: HttpRequest) -> HttpResponse:
     :param request: The HTTP request object.
     :return: An HTTP response with the Excel file attachment.
     """
+    print("Starting export")
+    timestamp_debug = pd.Timestamp.now()
     search = request.GET.get("search", "")
     search_column = request.GET.get("search_column", "")
     search_fields = (
@@ -298,8 +300,10 @@ def export_search_results(request: HttpRequest) -> HttpResponse:
 
     # Include all fields in the DataFrame, even if they are not displayed
     data_dicts = [row.__dict__ for row in rows]
+    print(f"Time to get data: {pd.Timestamp.now() - timestamp_debug}")
     # Add, remove, and reorder fields as needed
     export_df = format_data_for_export(data_dicts)
+    print(f"Time to format data: {pd.Timestamp.now() - timestamp_debug}")
 
     filename_base = "FTVA_DL_search_results"
     timestamp = pd.Timestamp.now().strftime("%Y%m%d_%H%M%S")
@@ -319,6 +323,8 @@ def export_search_results(request: HttpRequest) -> HttpResponse:
     buffer.seek(0)
     # Write the buffer content to the response
     response.write(buffer.read())
+    print("Returning response")
+    print(f"Total time: {pd.Timestamp.now() - timestamp_debug}")
 
     return response