Implements [SYS-1891](https://uclalibrary.atlassian.net/browse/SYS-1891)

Makes two performance-focused changes to the Export functionality of the FTVA Digital Data app:

1. Refactors the helper function `format_data_for_export()` in `views_utils.py` to avoid calling `SheetImport.objects.get()` for every item in the output set. Instead, `prefetch_related` and `select_related` are used to only query the database once, and lookup dictionaries are used to minimize the logic that happens inside the main loop.

This change is responsible for most of the performance improvement I saw locally, bringing the time to export all records from 42 seconds to 22 seconds.

2. Returns a CSV file via a `StreamingHttpResponse` instead of a binary Excel file with a standard `HttpResponse`. 

This change bought me an additional 5 seconds, taking the time to export all records down to 17 seconds.

`pandas` is still used internally to create the CSV file. I did not see any performance improvement if this is done with the `csv` package instead, and `pandas` allows for simpler code than would be needed with `csv.Dictwriter()`.

Additionally includes:

* Update to search results template to change button wording to "Export to CSV"
* Update to `requirements.txt` to remove `xlsxwriter` package
* Update to tests to work with new dict structure returned by `format_data_for_export()`
* Tag bump for deployment



[SYS-1891]: https://uclalibrary.atlassian.net/browse/SYS-1891?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ